sendmail: Add filetype dependent mailer args
Add a dict g:VimMailArgsByFiletype that can store arguments to the
mailer depending on the filetype. Use "-a" as default argument if
current filetype is no key in the dict.

By default, use the argument "-H" for the mail filetype, which makes
mutt and neomutt use the composed file as a draft file rather than
attaching it, what "-a" would do.